format: provide the "simplify" feature
Given that `gofumpt` enforces `gofmt`'s simplify flag, it's weird that
it's not applied via the Go API. Pure Go users like gopls are getting a
different experience, so this should be fixed.

This patch adds the support of the `simplify` feature via the Go API.

Fixes https://github.com/mvdan/gofumpt/issues/115
